Rename project-geometry-edit to project-geometries
