+++ app/templates/components/project-geometries.hbs
@@ -47,7 +47,7 @@
       {{/map.labs-layers}}
 
       {{!-- invoke the component for the specified draw mode --}}
-      {{component (concat 'project-geometry-edit/' type)
+      {{component (concat 'project-geometries/' type)
         model=model
         mode=mode
         map=map}}
