+++ app/templates/components/project-geometries/development-site.hbs
@@ -1,4 +1,4 @@
-{{#project-geometry-edit/draw-lots-to-union
+{{#project-geometries/draw-lots-to-union
   mode=mode
   map=map
   geometricProperty=model.developmentSite as |selection|
@@ -7,7 +7,7 @@
     handleClick=(action 'save')}}
     Save Development Site
   {{/selection.button}}
-{{/project-geometry-edit/draw-lots-to-union}}
+{{/project-geometries/draw-lots-to-union}}
 
 {{#ember-wormhole to='geometry-type-draw-explainer'}}
   Explainer Text here
