+++ app/templates/components/project-geometries/underlying-zoning.hbs
@@ -1,5 +1,5 @@
 {{#if model.underlyingZoning}}
-  {{#project-geometry-edit/draw-lots-to-union
+  {{#project-geometries/draw-lots-to-union
     mode=mode
     map=map
     geometricProperty=model.underlyingZoning as |selection|
@@ -8,7 +8,7 @@
       handleClick=(action 'save')}}
       Save Zoning District
     {{/selection.button}}
-  {{/project-geometry-edit/draw-lots-to-union}}
+  {{/project-geometries/draw-lots-to-union}}
 
   {{#map.source options=(mapbox-geojson-source model.underlyingZoning) as |source|}}
     {{source.layer
