+++ app/templates/projects/edit/geometry-edit.hbs
@@ -1,4 +1,4 @@
-{{project-geometry-edit
+{{project-geometries
   model=model
   type=type
   mode=mode}}
