+++ tests/integration/components/project-geometry-edit-test.js
@@ -5,22 +5,22 @@
 // import { render } from '@ember/test-helpers';
 // import hbs from 'htmlbars-inline-precompile';
 
-// module('Integration | Component | project-geometry-edit', function(hooks) {
+// module('Integration | Component | project-geometries', function(hooks) {
 //   setupRenderingTest(hooks);
 
 //   test('it renders', async function(assert) {
 //     // Set any properties with this.set('myProperty', 'value');
 //     // Handle any actions with this.set('myAction', function(val) { ... });
 
-//     await render(hbs`{{project-geometry-edit}}`);
+//     await render(hbs`{{project-geometries}}`);
 
 //     assert.equal(this.element.textContent.trim(), '');
 
 //     // Template block usage:
 //     await render(hbs`
-//       {{#project-geometry-edit}}
+//       {{#project-geometries}}
 //         template block text
-//       {{/project-geometry-edit}}
+//       {{/project-geometries}}
 //     `);
 
 //     assert.equal(this.element.textContent.trim(), 'template block text');
