+++ tests/integration/components/project-geometry-edit/commercial-overlays-test.js
@@ -3,7 +3,7 @@ import { setupRenderingTest } from 'ember-qunit';
 import { render } from '@ember/test-helpers';
 import hbs from 'htmlbars-inline-precompile';
 
-module('Integration | Component | project-geometry-edit/commercial-overlays', function(hooks) {
+module('Integration | Component | project-geometries/commercial-overlays', function(hooks) {
   setupRenderingTest(hooks);
 
   test('it renders', async function(assert) {
@@ -12,7 +12,7 @@ module('Integration | Component | project-geometry-edit/commercial-overlays', fu
 
     await render(hbs`
       {{#mapbox-gl as |map|}} 
-        {{project-geometry-edit/commercial-overlays map=map}}
+        {{project-geometries/commercial-overlays map=map}}
       {{/mapbox-gl}}
     `);
