+++ tests/integration/components/project-geometry-edit/draw-lots-to-union/button-test.js
@@ -3,22 +3,22 @@ import { setupRenderingTest } from 'ember-qunit';
 import { render } from '@ember/test-helpers';
 import hbs from 'htmlbars-inline-precompile';
 
-module('Integration | Component | project-geometry-edit/draw-lots-to-union/button', function(hooks) {
+module('Integration | Component | project-geometries/draw-lots-to-union/button', function(hooks) {
   setupRenderingTest(hooks);
 
   test('it renders', async function(assert) {
     // Set any properties with this.set('myProperty', 'value');
     // Handle any actions with this.set('myAction', function(val) { ... });
 
-    await render(hbs`{{project-geometry-edit/draw-lots-to-union/button}}`);
+    await render(hbs`{{project-geometries/draw-lots-to-union/button}}`);
 
     assert.equal(this.element.textContent.trim(), '');
 
     // Template block usage:
     await render(hbs`
-      {{#project-geometry-edit/draw-lots-to-union/button}}
+      {{#project-geometries/draw-lots-to-union/button}}
         template block text
-      {{/project-geometry-edit/draw-lots-to-union/button}}
+      {{/project-geometries/draw-lots-to-union/button}}
     `);
 
     assert.equal(this.element.textContent.trim(), 'template block text');
