+++ tests/integration/components/project-geometry-edit/project-area-test.js
@@ -3,7 +3,7 @@ import { setupRenderingTest } from 'ember-qunit';
 import { render } from '@ember/test-helpers';
 import hbs from 'htmlbars-inline-precompile';
 
-module('Integration | Component | project-geometry-edit/project-area', function(hooks) {
+module('Integration | Component | project-geometries/project-area', function(hooks) {
   setupRenderingTest(hooks);
 
   test('it renders', async function(assert) {
@@ -12,7 +12,7 @@ module('Integration | Component | project-geometry-edit/project-area', function(
 
     await render(hbs`
       {{#mapbox-gl as |map|}} 
-        {{project-geometry-edit/project-area map=map}}
+        {{project-geometries/project-area map=map}}
       {{/mapbox-gl}}
     `);
