+++ tests/integration/components/project-geometry-edit/special-purpose-districts-test.js
@@ -3,7 +3,7 @@ import { setupRenderingTest } from 'ember-qunit';
 import { render } from '@ember/test-helpers';
 import hbs from 'htmlbars-inline-precompile';
 
-module('Integration | Component | project-geometry-edit/special-purpose-districts', function(hooks) {
+module('Integration | Component | project-geometries/special-purpose-districts', function(hooks) {
   setupRenderingTest(hooks);
 
   test('it renders', async function(assert) {
@@ -12,7 +12,7 @@ module('Integration | Component | project-geometry-edit/special-purpose-district
 
     await render(hbs`
       {{#mapbox-gl as |map|}} 
-        {{project-geometry-edit/special-purpose-districts map=map}}
+        {{project-geometries/special-purpose-districts map=map}}
       {{/mapbox-gl}}
     `);
